app/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
python-dotenv==1.0.0
python-multipart==0.0.6
redis==5.0.1
rq==1.16.0
openai-whisper==20231117
vosk==0.3.45
langdetect==1.0.9
regex==2023.12.25

# Optional: for local translation models
# transformers==4.36.2
# torch==2.1.2
# sentencepiece==0.1.99

# Logging and monitoring
structlog==24.1.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-httpx==0.29.0
pytest-cov==4.1.0
